+++ simba/core/data/preprocessing_simba.py
@@ -1,10 +1,10 @@
 import copy
 
+from simba.core.data.loader_saver import LoaderSaver
 from simba.core.data.loaders import LoadData
+from simba.core.data.preprocessor import Preprocessor
 from simba.core.data.spectrum import SpectrumExt
-from simba.loader_saver import LoaderSaver
-from simba.logger_setup import logger
-from simba.preprocessor import Preprocessor
+from simba.utils.logger_setup import logger
 
 
 class PreprocessingSimba: